+++ packages/jaspr/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `table` and related html methods.
 - Added `value` parameter to `select()` method.
+- Add 'standalone' option to `renderComponent` method.
 
 ## 0.16.2